e2e-tests reward points: moved auxiliary functions to e2e-tests/src/rewards.rs
