+++ e2e-tests/src/lib.rs
@@ -4,5 +4,6 @@ pub use config::Config;
 mod accounts;
 mod cases;
 mod config;
+mod rewards;
 mod test;
 mod transfer;
